refactor(tabs): Apply new compose contract convention.
